requirements.txt
django==1.11
django-bootstrap3
gunicorn
django-widget-tweaks
django-tinymce
python-magic